Refactor test_renew_common() - Fix #41
This move all the tests in main_test.py, that calls on test_renew_common(), into
renewal_test.py

* adds imports and encoding to renewal_test.py

* disable pylint check for amount of public methods

* copy dummy_renewal_config method to renewal_test.py

* move renew_common to renewal_test.py

* move all tests using renew_common to renewal_test.py

* removes unused imports from main_test.py

* removes moved tests from main_test.py
